+++ samples/api.callback.js
@@ -2,7 +2,9 @@
 const stringify = require('../lib')
 const assert = require('assert')
 
-input = [ [ '1', '2', '3', '4' ], [ 'a', 'b', 'c', 'd' ] ]
-stringify(input, function(err, output){
+stringify([
+  [ '1', '2', '3', '4' ],
+  [ 'a', 'b', 'c', 'd' ]
+], function(err, output){
   assert.equal(output, '1,2,3,4\na,b,c,d\n')
 })